A first working version of the base line AI implemented, it can usually clear about 5-20k rows

+++ src/main/java/karlney/tetris/ai/IETetrisBoardEvaluator.java
@@ -49,7 +49,7 @@ public class IETetrisBoardEvaluator implements BoardEvaluator{
 
     @Override
     public double score(Board board, Piece piece, int rows) {
-        int lh = BoardMeasuresUtil.getLandingHeight(board.getRows(),piece);
+        double lh = BoardMeasuresUtil.getLandingHeight(board.getRows(),piece);
         int re = rows;
         int rt = BoardMeasuresUtil.getRowTransitions(board);
         int ct = BoardMeasuresUtil.getColumnTransitions(board);
